+++ resources/META-INF/plugin.xml
@@ -56,8 +56,8 @@
 
   <extensions defaultExtensionNs="com.intellij">
     <!-- Add your extensions here -->
-    <projectService serviceImplementation="com.cleverCloud.cleverIdea.ProjectSettings"/>
-    <applicationService serviceImplementation="com.cleverCloud.cleverIdea.ApplicationSettings"/>
+    <projectService serviceImplementation="com.cleverCloud.cleverIdea.settings.ProjectSettings"/>
+    <applicationService serviceImplementation="com.cleverCloud.cleverIdea.settings.ApplicationSettings"/>
     <checkoutProvider implementation="com.cleverCloud.cleverIdea.vcs.CleverCheckoutProvider"/>
   </extensions>
 
@@ -68,7 +68,7 @@
   <project-components>
     <!-- Add your project components here -->
     <component>
-      <implementation-class>com.cleverCloud.cleverIdea.CleverCloudProjectComponent</implementation-class>
+      <implementation-class>com.cleverCloud.cleverIdea.component.CleverCloudProjectComponent</implementation-class>
     </component>
   </project-components>
